Use the backing arrays.

+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/AbstractLongStorage.java
@@ -273,7 +273,7 @@ public abstract class AbstractLongStorage extends Storage<Long> implements Colum
     int size = (int) parent.getSize();
     return new ComputedNullableLongStorage(size + count) {
       @Override
-      protected Long computeItem(int idx) {
+      protected Long computeItem(long idx) {
         if (idx < size) {
           return parent.getItemBoxed(idx);
         } else {
