+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/ComputedNullableLongStorage.java
@@ -14,7 +14,7 @@ import org.graalvm.polyglot.Context;
  */
 public abstract class ComputedNullableLongStorage extends AbstractLongStorage
     implements ColumnStorageWithNothingMap {
-  protected abstract Long computeItem(int idx);
+  protected abstract Long computeItem(long idx);
 
   private BitSet isNothing;
