+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/DoubleStorage.java
@@ -397,7 +397,7 @@ public final class DoubleStorage extends Storage<Double>
     ComputedNullableLongStorage longAdapter =
         new ComputedNullableLongStorage(size) {
           @Override
-          protected Long computeItem(int idx) {
+          protected Long computeItem(long idx) {
             if (parent.isNothing(idx)) {
               return null;
             }
